requirements.txt
# Alpaca SDK for stock trading and market data
alpaca-py>=0.21.0

# Yahoo Finance for float data
yfinance>=0.2.30

# Testing
pytest>=7.0.0
pytest-cov>=4.0.0
freezegun>=1.2.0

# Environment variable loading
python-dotenv>=1.0.0

# Data analysis
pandas>=2.0.0

# YAML config parsing
pyyaml>=6.0

# Timezone handling
pytz>=2023.3

# LLM news analysis (optional - falls back to V1 stub without API key)
anthropic>=0.42.0

# Telegram notifications
aiohttp>=3.9.0
